requirements.txt
numpy
opencv-python
pygame
tensorflow
keras
paho-mqtt
